requirements.txt
streamlit
scikit-learn==1.1.2
pillow